Baseline implementation of session capture.
For now we simply grab the raw h264 feed from the encoder and manually
save the NALUs into a save file. This approach is only a proof of
concept and will be improved for production.

This commit also changes the minimum wait time between frames(for some
reason capped at 40ms, a.k.a. 25 fps) down to 16ms(up to 60fps).